updating class names
fixing an issue with pytest: everytime pytest runs, it runs all the classes whose name starts with "Test". In the previous code not all the classes were meant to be tested but they were supposed to simulate a signal or a widget. The names of these classes were changed from "TestClass" to "FakeClass"