Added dockable widget to project_window
Added a dock_view_ui.py file, which is responsible for the dock_widget style (now it's missing only the menubar on the top). Also modified the windows_factory to inherit the window style from the dock_view_ui.py.
In the end, small size modifications and changed the way to open secondary windows with a QDockWidget