add theme + dockable windows + handle layouts
Inside pzero.py we set the theme and the icon, we also have created a button inside project_window (which is the main window) for switching from one theme to another one (and for this we created the buttons inside the menu of project_window).
We created two other buttons for saving and restoring a window layout always inside the help menu inside project_window.py.
Another important new feature is the dockable windows, now all the secondary windows can be docked and undocked from the main window. We still need to adjust the resize of the components inside the secondary windows.